Move incoming sync address helper to database utilities
